feat!: migrate to ESM
BREAKING CHANGES: ESM and node 18 minimum
