+++ .lintstagedrc.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+  '**/*.{js,json,md}?(x)': () => 'npm run reformat',
+};
