--- .lintstagedrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  '**/*.{js,json,md}?(x)': () => 'npm run reformat'
-};
