+++ .mocharc.json
@@ -1,9 +1,10 @@
 {
-  "require": "ts-node/register,source-map-support/register",
+  "require": ["ts-node/register"],
   "watch-extensions": "ts",
   "extensions": [".ts", ".js", ".json"],
   "watch-files": ["src", "test"],
   "recursive": true,
   "reporter": "spec",
-  "timeout": 5000
+  "timeout": 5000,
+  "node-option": ["loader=ts-node/esm"]
 }
