--- bin/run
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-require('@oclif/core').run().then(require('@oclif/core/flush')).catch(require('@oclif/core/handle'));
