+++ bin/run.js
@@ -1,0 +1,9 @@
+#!/usr/bin/env node
+
+// eslint-disable-next-line node/shebang
+async function main() {
+  const { execute } = await import('@oclif/core');
+  await execute({ dir: import.meta.url });
+}
+
+await main();
