+++ messages/messages.md
@@ -1,3 +1,7 @@
 # errorWithOrg
 
 Org: %s: %s
+
+# noAuthFound
+
+No authenticated orgs found.
