+++ package.json
@@ -1,63 +1,40 @@
 {
   "name": "@salesforce/plugin-signups",
   "description": "Commands to interact with org shapes",
-  "version": "1.5.6",
+  "version": "2.0.0",
   "author": "Salesforce",
-  "main": "lib/index.js",
   "bugs": "https://github.com/forcedotcom/cli/issues",
   "dependencies": {
-    "@oclif/core": "^2.15.0",
+    "@oclif/core": "^3.11.0",
     "@salesforce/core": "^5.3.14",
     "@salesforce/kit": "^3.0.13",
-    "@salesforce/sf-plugins-core": "^3.1.28",
-    "chalk": "^4.1.2",
-    "change-case": "^4.1.2",
-    "tslib": "^2"
+    "@salesforce/sf-plugins-core": "^4.1.1",
+    "chalk": "^5.3.0",
+    "change-case": "^4.1.2"
   },
   "devDependencies": {
-    "@oclif/plugin-command-snapshot": "^4.0.16",
-    "@salesforce/cli-plugins-testkit": "^4.4.7",
-    "@salesforce/dev-config": "^4.1.0",
-    "@salesforce/dev-scripts": "^5.12.2",
+    "@oclif/plugin-command-snapshot": "^5.0.2",
+    "@salesforce/cli-plugins-testkit": "^5.0.4",
+    "@salesforce/dev-scripts": "^6.0.4",
     "@salesforce/plugin-command-reference": "^3.0.44",
-    "@salesforce/prettier-config": "^0.0.3",
     "@salesforce/ts-sinon": "^1.4.19",
     "@salesforce/ts-types": "^2.0.6",
-    "@swc/core": "^1.3.39",
     "@types/chai-as-promised": "^7.1.7",
     "@types/chai-string": "^1.4.3",
     "@types/inquirer": "^9.0.6",
     "@types/shelljs": "^0.8.14",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
-    "chai": "^4.3.10",
     "chai-as-promised": "7.1.1",
     "chai-string": "^1.5.0",
-    "eslint": "^8.53.0",
-    "eslint-config-prettier": "^8.10.0",
-    "eslint-config-salesforce": "^2.0.2",
-    "eslint-config-salesforce-license": "^0.2.0",
-    "eslint-config-salesforce-typescript": "^2.0.0",
-    "eslint-plugin-header": "^3.1.1",
-    "eslint-plugin-import": "^2.29.0",
-    "eslint-plugin-jsdoc": "^46.8.2",
     "eslint-plugin-sf-plugin": "^1.16.12",
-    "husky": "^7.0.4",
     "jsforce": "^2.0.0-beta.27",
-    "mocha": "^9.1.3",
-    "nyc": "^15.1.0",
-    "oclif": "^3.17.2",
-    "prettier": "^2.8.8",
-    "pretty-quick": "^3.1.0",
+    "oclif": "^4.0.3",
     "shx": "0.3.4",
-    "sinon": "^11.1.1",
-    "ts-node": "^10.2.1",
-    "typescript": "^5.2.2",
-    "wireit": "^0.14.1"
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   },
   "config": {},
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=18.0.0"
   },
   "files": [
     "/lib",
@@ -69,14 +46,16 @@
   "keywords": [
     "force",
     "salesforce",
-    "sfdx",
     "salesforcedx",
+    "sf",
+    "sf-plugin",
+    "sfdx",
     "sfdx-plugin"
   ],
   "license": "BSD-3-Clause",
   "oclif": {
     "commands": "./lib/commands",
-    "bin": "sfdx",
+    "bin": "sf",
     "devPlugins": [
       "@oclif/plugin-command-snapshot",
       "@salesforce/plugin-command-reference"
@@ -136,19 +115,21 @@
           }
         }
       }
-    }
+    },
+    "flexibleTaxonomy": true,
+    "topicSeparator": " "
   },
   "repository": "salesforcecli/plugin-signups",
   "scripts": {
     "build": "wireit",
     "clean": "sf-clean",
     "clean-all": "sf-clean all",
-    "clean:lib": "shx rm -rf lib && shx rm -rf coverage && shx rm -rf .nyc_output && shx rm -f oclif.manifest.json",
+    "clean:lib": "shx rm -rf lib && shx rm -rf coverage && shx rm -rf .nyc_output && shx rm -f oclif.manifest.json oclif.lock",
     "compile": "wireit",
     "docs": "sf-docs",
     "format": "wireit",
     "lint": "wireit",
-    "postpack": "shx rm -f oclif.manifest.json",
+    "postpack": "shx rm -f oclif.manifest.json oclif.lock",
     "prepack": "sf-prepack",
     "prepare": "sf-install",
     "test": "wireit",
@@ -236,7 +217,7 @@
       "output": []
     },
     "test:command-reference": {
-      "command": "\"./bin/dev\" commandreference:generate --erroronwarnings",
+      "command": "ts-node \"./bin/dev.js\" commandreference:generate --erroronwarnings",
       "files": [
         "src/**/*.ts",
         "messages/**",
@@ -247,7 +228,7 @@
       ]
     },
     "test:deprecation-policy": {
-      "command": "\"./bin/dev\" snapshot:compare",
+      "command": "ts-node \"./bin/dev.js\" snapshot:compare",
       "files": [
         "src/**/*.ts"
       ],
@@ -257,12 +238,14 @@
       ]
     },
     "test:json-schema": {
-      "command": "\"./bin/dev\" schema:compare",
+      "command": "ts-node \"./bin/dev.js\" schema:compare",
       "files": [
         "src/**/*.ts",
         "schemas"
       ],
       "output": []
     }
-  }
+  },
+  "exports": "./lib/index.js",
+  "type": "module"
 }
