+++ src/commands/org/create/shape.ts
@@ -5,6 +5,8 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import {
   SfCommand,
   requiredOrgFlagWithDeprecations,
@@ -13,9 +15,9 @@ import {
 } from '@salesforce/sf-plugins-core';
 import { Messages, Connection, Logger } from '@salesforce/core';
 import type { SaveResult } from 'jsforce';
-import { isShapeEnabled, JsForceError } from '../../../shared/orgShapeListUtils';
+import { isShapeEnabled, JsForceError } from '../../../shared/orgShapeListUtils.js';
 
-Messages.importMessagesDirectory(__dirname);
+Messages.importMessagesDirectory(dirname(fileURLToPath(import.meta.url)));
 const messages = Messages.loadMessages('@salesforce/plugin-signups', 'shape.create');
 
 export interface ShapeCreateResult {
