+++ src/commands/org/delete/shape.ts
@@ -5,6 +5,8 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import {
   Flags,
   SfCommand,
@@ -12,31 +14,12 @@ import {
   orgApiVersionFlagWithDeprecations,
   loglevel,
 } from '@salesforce/sf-plugins-core';
-import { Messages, Connection, SfError } from '@salesforce/core';
-import { isShapeEnabled, JsForceError } from '../../../shared/orgShapeListUtils';
+import { Messages, SfError } from '@salesforce/core';
+import { isShapeEnabled } from '../../../shared/orgShapeListUtils.js';
+import utils, { DeleteAllResult } from '../../../shared/deleteUtils.js';
 
-Messages.importMessagesDirectory(__dirname);
+Messages.importMessagesDirectory(dirname(fileURLToPath(import.meta.url)));
 const messages = Messages.loadMessages('@salesforce/plugin-signups', 'shape.delete');
-
-interface ShapeRepresentation {
-  Id: string;
-  Status: string;
-  Edition: string;
-  Features: string;
-  Settings: string;
-  CreatedDate: string;
-  Description: string;
-}
-
-type FailureMsg = {
-  shapeId: string;
-  message: string;
-};
-
-type DeleteAllResult = {
-  shapeIds: string[];
-  failures: FailureMsg[];
-};
 
 export interface OrgShapeDeleteResult extends DeleteAllResult {
   orgId: string;
@@ -78,7 +61,7 @@ export class OrgShapeDeleteCommand extends SfCommand<OrgShapeDeleteResult | unde
       throw messages.createError('noAccess', [username]);
     }
 
-    const deleteRes = await deleteAll(conn, username);
+    const deleteRes = await utils.deleteAll(conn, username);
 
     if (deleteRes.shapeIds.length === 0) {
       this.info(messages.getMessage('noShapesHumanSuccess', [orgId]));
@@ -112,51 +95,6 @@ export class OrgShapeDeleteCommand extends SfCommand<OrgShapeDeleteResult | unde
   }
 }
 
-/**
- * Delete all ShapeRepresentation records for the shapeOrg.
- *
- * @return List of SR IDs that were deleted
- */
-export const deleteAll = async (conn: Connection, username: string): Promise<DeleteAllResult> => {
-  let shapeIds: string[] = [];
-  const deleteAllResult: DeleteAllResult = {
-    shapeIds: [],
-    failures: [],
-  };
-  try {
-    const result = await conn.query<ShapeRepresentation>('SELECT Id FROM ShapeRepresentation');
-    if (result.totalSize === 0) {
-      return deleteAllResult;
-    }
-    shapeIds = result.records.map((shape) => shape.Id);
-  } catch (err) {
-    const JsForceErr = err as JsForceError;
-    if (JsForceErr.errorCode && JsForceErr.errorCode === 'INVALID_TYPE') {
-      // ShapeExportPref is not enabled, or user does not have CRUD access
-      throw messages.createError('noAccess', [username]);
-    }
-    // non-access error
-    throw JsForceErr;
-  }
-
-  await Promise.all(
-    shapeIds.map(async (id) => {
-      try {
-        const delResult = await conn.sobject('ShapeRepresentation').delete(id);
-        if (delResult.success) {
-          deleteAllResult.shapeIds.push(id);
-        }
-      } catch (err) {
-        deleteAllResult.failures.push({
-          shapeId: id,
-          message: err instanceof Error ? err.message : 'error contained no message',
-        });
-      }
-    })
-  );
-
-  return deleteAllResult;
-};
 const setExitCode = (code: number): void => {
   process.exitCode = code;
 };
