+++ src/commands/org/list/snapshot.ts
@@ -4,6 +4,8 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+import { dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import {
   loglevel,
   orgApiVersionFlagWithDeprecations,
@@ -11,9 +13,9 @@ import {
   SfCommand,
 } from '@salesforce/sf-plugins-core';
 import { Messages } from '@salesforce/core';
-import { OrgSnapshot, queryAll, printRecordTable } from '../../../shared/snapshot';
+import { OrgSnapshot, queryAll, printRecordTable } from '../../../shared/snapshot.js';
 
-Messages.importMessagesDirectory(__dirname);
+Messages.importMessagesDirectory(dirname(fileURLToPath(import.meta.url)));
 const messages = Messages.loadMessages('@salesforce/plugin-signups', 'snapshot.list');
 
 export class SnapshotList extends SfCommand<OrgSnapshot[]> {
