+++ src/shared/snapshot.ts
@@ -5,11 +5,13 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import { ux } from '@oclif/core';
 import { Connection, SfError, Messages } from '@salesforce/core';
 import { capitalCase } from 'change-case';
 
-Messages.importMessagesDirectory(__dirname);
+Messages.importMessagesDirectory(dirname(fileURLToPath(import.meta.url)));
 const messages = Messages.loadMessages('@salesforce/plugin-signups', 'snapshot');
 
 export interface OrgSnapshotRequest {
