+++ test/shape/delete.test.ts
@@ -5,19 +5,20 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { resolve } from 'node:path';
+import { dirname, resolve } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import { Config } from '@oclif/core';
 import { use, expect, config as chaiConfig } from 'chai';
-import * as chaiAsPromised from 'chai-as-promised';
-import { TestContext, MockTestOrgData } from '@salesforce/core/lib/testSetup';
+import chaiAsPromised from 'chai-as-promised';
+import { TestContext, MockTestOrgData } from '@salesforce/core/lib/testSetup.js';
 
 import { SfCommand } from '@salesforce/sf-plugins-core';
 import type { SaveResult } from 'jsforce';
-import * as sinon from 'sinon';
+import sinon from 'sinon';
 import { ensureJsonMap, ensureString, AnyJson } from '@salesforce/ts-types';
-import { OrgShapeDeleteCommand } from '../../src/commands/org/delete/shape';
-import * as deleteFunctions from '../../src/commands/org/delete/shape';
-import { queryShapeEnabledResponse } from '../shared/apiResponses';
+import { OrgShapeDeleteCommand } from '../../src/commands/org/delete/shape.js';
+import utils from '../../src/shared/deleteUtils.js';
+import { queryShapeEnabledResponse } from '../shared/apiResponses.js';
 
 use(chaiAsPromised);
 chaiConfig.truncateThreshold = 0;
@@ -25,7 +26,7 @@ chaiConfig.truncateThreshold = 0;
 describe('org:shape:delete', () => {
   const $$ = new TestContext();
   const testOrg = new MockTestOrgData();
-  const config = new Config({ root: resolve(__dirname, '../../package.json') });
+  const config = new Config({ root: resolve(dirname(fileURLToPath(import.meta.url)), '../../package.json') });
 
   const sandbox = sinon.createSandbox();
 
@@ -113,7 +114,7 @@ describe('org:shape:delete', () => {
       throw new Error('Unexpected request: ' + JSON.stringify(request));
     };
 
-    sandbox.stub(deleteFunctions, 'deleteAll').resolves({
+    sandbox.stub(utils, 'deleteAll').resolves({
       shapeIds: ['3SR000000000123'],
       failures: [{ shapeId: '3SR000000000124', message: 'MALFORMED ID' }],
     });
@@ -133,7 +134,7 @@ describe('org:shape:delete', () => {
   });
 
   it('no shapes', async () => {
-    sandbox.stub(deleteFunctions, 'deleteAll').resolves({
+    sandbox.stub(utils, 'deleteAll').resolves({
       shapeIds: [],
       failures: [],
     });
