+++ test/shared/orgShapeListUtils.test.ts
@@ -5,14 +5,14 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { expect } from 'chai';
-import * as chai from 'chai';
-import * as chaiAsPromised from 'chai-as-promised';
+import chai from 'chai';
+import chaiAsPromised from 'chai-as-promised';
 
 chai.use(chaiAsPromised);
-import { TestContext, MockTestOrgData } from '@salesforce/core/lib/testSetup';
+import { TestContext, MockTestOrgData } from '@salesforce/core/lib/testSetup.js';
 import { AnyJson, ensureJsonMap, ensureString } from '@salesforce/ts-types';
 
-import { getAllShapesFromOrg } from '../../src/shared/orgShapeListUtils';
+import { getAllShapesFromOrg } from '../../src/shared/orgShapeListUtils.js';
 
 describe('shape list utils', () => {
   const $$ = new TestContext();
