+++ test/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "@salesforce/dev-config/tsconfig-test-strict",
+  "extends": "@salesforce/dev-config/tsconfig-test-strict-esm",
   "include": ["./**/*.ts"],
   "compilerOptions": {
     "strictNullChecks": true,
